+++ ios/help_scout_beacon/Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version: 5.9
+// swift-tools-version: 6.2
 import PackageDescription
 
 let package = Package(
@@ -20,5 +20,6 @@ let package = Package(
             ],
             path: "Sources/help_scout_beacon"
         ),
-    ]
+    ],
+    swiftLanguageModes: [.v6]
 )